This commit makes the random increments on `Monotonic` reads to always
be computed using bytes read from the underlying entropy source, whereas
before they'd be always derived from a `rand.Reader`.

In the process, I've refactored the benchmarks to be more dry using a table driven test approach. While this had the desired effect, it also affected the benchmark numbers slightly.

So here's a `benchstat` output of the comparison of the new and old code, but with old benchmark test code still.

```
name                                                old time/op   new time/op   delta
New/WithMonotonicEntropy_SameTimestamp_Inc0-8        73.4ns ± 1%   83.6ns ± 2%  +13.89%  (p=0.000 n=10+10)
New/WithMonotonicEntropy_DifferentTimestamp_Inc0-8   82.4ns ± 2%   82.7ns ± 2%     ~     (p=0.424 n=10+10)
New/WithMonotonicEntropy_SameTimestamp_Inc1-8        42.4ns ± 1%   45.2ns ± 3%   +6.72%  (p=0.000 n=10+10)
New/WithMonotonicEntropy_DifferentTimestamp_Inc1-8   83.3ns ± 1%   84.3ns ± 4%     ~     (p=0.175 n=9+10)

name                                                old speed     new speed     delta
New/WithMonotonicEntropy_SameTimestamp_Inc0-8       218MB/s ± 1%  192MB/s ± 1%  -12.04%  (p=0.000 n=10+9)
New/WithMonotonicEntropy_DifferentTimestamp_Inc0-8  194MB/s ± 2%  194MB/s ± 2%     ~     (p=0.436 n=10+10)
New/WithMonotonicEntropy_SameTimestamp_Inc1-8       378MB/s ± 1%  354MB/s ± 3%   -6.27%  (p=0.000 n=10+10)
New/WithMonotonicEntropy_DifferentTimestamp_Inc1-8  192MB/s ± 1%  190MB/s ± 4%     ~     (p=0.190 n=10+10)
```

I think this is acceptable, but please let me know if you see opportunities for optimisation.