@tsenart 2018-10-02
> Coverage now at 95% :o

Thanks for bringing it up! I caught and fixed bugs by getting the coverage up again.